Add get User method and listeners cleanup
Fixed a memory leak issue on rouge listeners, as listeners are not automatically removed on page change, we need to remove them manually.

On any .onSnapshot methods, please push that to window.removeFirebaseListener Array and it will be removed automatically on page change. For eg see the closet.js file.

@Aleirbag04 @ernestoguevara92 @jhajj01